centralized validation logic
