+++ resources/views/admin/appointments/create.blade.php
@@ -33,5 +33,5 @@
 
 
 @section('scripts')
-    @vite(['resources/js/commons/validate'])
+    @vite(['resources/js/validations/appointment-form'])
 @endsection
